+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Page is now a state instead of being defined by `from` and `too`
+
 ## [0.4.0] - 2019-12-27
 
 ### Added